production_requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
openpyxl==3.1.2
Pillow==10.1.0
Werkzeug==2.3.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
msal==1.24.1
aiofiles==23.2.1
httpx==0.25.2
openai==1.3.5
python-docx==0.8.11
lxml==4.9.3
